pyproject.toml
[project]
name = "watch-number-generator"
version = "0.1.0"
description = "Generate 3D printable numbers for watch dials"
authors = [{name = "Alberto Franzin", email = "franzinalberto01@gmail.com"}]
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygobject>=3.42.0",
    "numpy>=1.24.0",
    "numpy-stl>=3.0.0",
    "trimesh>=4.0.0",
    "pillow>=10.0.0",
    "pycairo>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
